NIAD-1880: Move general properties to separate class, refactor package name
