+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/config/GeneralProperties.java
@@ -1,0 +1,15 @@
+package uk.nhs.adaptors.pss.translator.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Component
+@ConfigurationProperties(prefix = "general")
+@Getter
+@Setter
+public class GeneralProperties {
+    private String fromOdsCode;
+}
