+++ gp2gp-translator/src/test/java/uk/nhs/adaptors/pss/translator/mapper/LocationMapperTest.java
@@ -1,15 +1,17 @@
 package uk.nhs.adaptors.pss.translator.mapper;
 
-import lombok.SneakyThrows;
-import org.hl7.fhir.dstu3.model.Location;
-import org.hl7.fhir.dstu3.model.Identifier;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.springframework.util.ResourceUtils.getFile;
+
+import static uk.nhs.adaptors.pss.translator.testutil.XmlUnmarshallUtil.unmarshallFile;
+
 import org.hl7.fhir.dstu3.model.ContactPoint;
+import org.hl7.fhir.dstu3.model.Identifier;
+import org.hl7.fhir.dstu3.model.Location;
 import org.hl7.v3.RCMRMT030101UK04EhrComposition;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.util.ResourceUtils.getFile;
-import static uk.nhs.adaptors.pss.translator.util.XmlUnmarshallUtil.unmarshallFile;
+import lombok.SneakyThrows;
 
 public class LocationMapperTest {
     private static final String XML_RESOURCES_BASE = "XML/Location/";
