+++ gp2gp-translator/src/test/java/uk/nhs/adaptors/pss/translator/util/ParameterUtilsTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.Test;
 import uk.nhs.adaptors.pss.translator.testutil.CreateParametersUtil;
 
 public class ParameterUtilsTest {
-
     private static final String TEST_NHS_NUMBER = "TEST_NHS_NUMBER";
 
     @Test
@@ -16,5 +15,4 @@ public class ParameterUtilsTest {
         String nhsNumber = ParametersUtils.getNhsNumberFromParameters(parameters).get().getValue();
         assertEquals(TEST_NHS_NUMBER, nhsNumber);
     }
-
 }
