+++ gpc-api-facade/src/test/java/uk/nhs/adaptors/pss/gpc/testutil/CreateParametersUtil.java
@@ -42,5 +42,4 @@ public class CreateParametersUtil {
 
         return nhsNumberComponent;
     }
-
 }
